feat: Update addFlight and updateFlight methods in FlightController
This commit updates the `addFlight` and `updateFlight` methods in the `FlightController` class. The methods now accept an additional parameter `HttpServletRequest request` to retrieve the `userId` attribute. This allows for retrieving the `Airline` entity using the `userId` instead of the `airlineId` from the `FlightDTO`. This change ensures that the correct `Airline` is associated with the `Flight` entity.